requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
jsonschema
aiohttp
motor
pytest-asyncio
tenacity
mongomock-motor
pytest-timeout
pytest-mock
